refactored recipeCard
