+++ src/components/recipeDisplay/RecipeContainer.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import RecipesBrief from './RecipeBrief';
+import RecipeCard from './RecipeCard';
 import './RecipeContainer.css';
 
 export default function RecipeContainer({ recipes }) {
   return (
     <div className="recipe-brief-container mx-auto w-full max-w-screen-xl">
-      <RecipesBrief recipes={recipes} />
+      <RecipeCard recipes={recipes} />
     </div>
   );
 }
